requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Authentication
supabase==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cloud Storage
google-cloud-storage==2.18.2

# AI/ML
openai==1.54.4
pymilvus==2.4.8
numpy==1.26.4
pillow==10.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
ruff==0.7.1
mypy==1.13.0
types-python-jose==3.3.4.20240106